Organize existing dashboard template into the reusable components
Removing unnused code and setup react router and redux
