+++ portal-ui/src/screens/SignupPage/index.tsx
@@ -14,12 +14,13 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-
 import React from 'react';
 import Avatar from '@material-ui/core/Avatar';
 import Button from '@material-ui/core/Button';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import TextField from '@material-ui/core/TextField';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import Checkbox from '@material-ui/core/Checkbox';
 import Link from '@material-ui/core/Link';
 import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
@@ -27,10 +28,8 @@ import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
 import Typography from '@material-ui/core/Typography';
 import {makeStyles} from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
-import Copyright from "../copyright";
-import request from "superagent";
-import storage from "local-storage-fallback";
-import {useHistory} from "react-router";
+import Copyright from "../../components/Copyright";
+
 
 const useStyles = makeStyles(theme => ({
     '@global': {
@@ -55,40 +54,9 @@ const useStyles = makeStyles(theme => ({
     submit: {
         margin: theme.spacing(3, 0, 2),
     },
-    errorBlock: {
-        color: 'red',
-    }
 }));
-const Login: React.FC = () => {
+const Signup: React.FC = () => {
     const classes = useStyles();
-    const {push} = useHistory()
-    const [email, setEmail] = React.useState<string>('');
-    const [password, setPassword] = React.useState<string>('');
-    const [company, setCompany] = React.useState<string>('');
-    const [error, setError] = React.useState<string>('');
-
-    const formSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        const url = '/api/v1/users/login';
-        request
-            .post(url)
-            .send({email: email, password: password, company: company})
-            .then((res: any) => {
-                if (res.body.jwt_token) {
-                    // store the jwt token
-                    storage.setItem('token', res.body.jwt_token);
-                    return res.body.jwt_token;
-                } else if (res.body.error) {
-                    // throw will be moved to catch block once bad login returns 403
-                    throw res.body.error;
-                }
-            }).then(() => {
-            push('/dashboard');
-        })
-            .catch(err => {
-                setError(err);
-            });
-    }
 
     return (
         <Container component="main" maxWidth="xs">
@@ -98,30 +66,19 @@ const Login: React.FC = () => {
                     <LockOutlinedIcon/>
                 </Avatar>
                 <Typography component="h1" variant="h5">
-                    Acme Storage Login
+                    Acme Storage Sign up
                 </Typography>
-                <form className={classes.form} noValidate onSubmit={formSubmit}>
+                <form className={classes.form} noValidate>
                     <Grid container spacing={2}>
-                        {error !== '' && (
-                            <Grid item xs={12}>
-                                <Typography component="p" variant="body1" className={classes.errorBlock}>
-                                    {`${error}`}
-                                </Typography>
-                            </Grid>
-                        )}
                         <Grid item xs={12}>
                             <TextField
-                                autoComplete="company_name"
-                                name="company_name"
+                                autoComplete="fullname"
+                                name="fullname"
                                 variant="outlined"
                                 required
                                 fullWidth
-                                value={company}
-                                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                                    setCompany(e.target.value)
-                                }
-                                id="company_name"
-                                label="Company"
+                                id="fullname"
+                                label="Full Name"
                                 autoFocus
                             />
                         </Grid>
@@ -132,10 +89,6 @@ const Login: React.FC = () => {
                                 required
                                 fullWidth
                                 id="email"
-                                value={email}
-                                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                                    setEmail(e.target.value)
-                                }
                                 label="Email Address"
                                 name="email"
                                 autoComplete="email"
@@ -146,15 +99,83 @@ const Login: React.FC = () => {
                                 variant="outlined"
                                 required
                                 fullWidth
-                                value={password}
-                                onChange={(e: React.ChangeEvent<HTMLInputElement>) =>
-                                    setPassword(e.target.value)
-                                }
                                 name="password"
                                 label="Password"
                                 type="password"
                                 id="password"
                                 autoComplete="current-password"
+                            />
+                        </Grid>
+                        <hr/>
+                        <Grid item xs={12}>
+                            <TextField
+                                variant="outlined"
+                                required
+                                fullWidth
+                                id="company"
+                                label="Company Name"
+                                name="email"
+                                autoComplete="email"
+                            />
+                        </Grid>
+                        <Grid item xs={12}>
+                            <TextField
+                                variant="outlined"
+                                required
+                                fullWidth
+                                id="company_short_name"
+                                label="Short Name"
+                                name="company_short_name"
+                                autoComplete="company_short_name"
+                            />
+                        </Grid>
+
+                        <Grid item xs={12} sm={8}>
+                            <TextField
+                                variant="outlined"
+                                required
+                                fullWidth
+                                id="credit_card_number"
+                                label="Credit Card Number"
+                                name="credit_card_number"
+                            />
+                        </Grid>
+                        <Grid item xs={12} sm={4}>
+                            <TextField
+                                variant="outlined"
+                                required
+                                fullWidth
+                                id="ccv"
+                                label="CCV"
+                                name="ccv"
+                            />
+                        </Grid>
+                        <Grid item xs={12} sm={6}>
+                            <TextField
+                                variant="outlined"
+                                required
+                                fullWidth
+                                id="expiration_mm"
+                                label="Expiration MM"
+                                name="expiration_mm"
+                                autoComplete="expiration_mm"
+                            />
+                        </Grid>
+                        <Grid item xs={12} sm={6}>
+                            <TextField
+                                variant="outlined"
+                                required
+                                fullWidth
+                                id="expiration_yy"
+                                label="Expiration YY"
+                                name="expiration_YY"
+                                autoComplete="expiration_yy"
+                            />
+                        </Grid>
+                        <Grid item xs={12}>
+                            <FormControlLabel
+                                control={<Checkbox value="allowExtraEmails" color="primary"/>}
+                                label="Agree to Terms & Conditions."
                             />
                         </Grid>
                     </Grid>
@@ -165,12 +186,12 @@ const Login: React.FC = () => {
                         color="primary"
                         className={classes.submit}
                     >
-                        Login
+                        Sign Up
                     </Button>
                     <Grid container justify="flex-end">
                         <Grid item>
                             <Link href="#" variant="body2">
-                                Forgot Password?
+                                Already have an account? Sign in
                             </Link>
                         </Grid>
                     </Grid>
@@ -183,4 +204,4 @@ const Login: React.FC = () => {
     );
 };
 
-export default Login;
+export default Signup;
